refactor: three-tier device monitor fallback (rtnetlink → ip -j → ip text)
Split linux.rs into linux/{mod,rtnetlink,ip_json,ip_text}.rs.

Probe sequence: rtnetlink multicast → ip -j monitor (iproute2 ≥ 4.12) →
ip monitor text. Each tier is probed before use so the best available
backend is selected at runtime.

ip_json parser dispatches on field presence (no event type field in
iproute2 JSON): "ifname" → link event, "dev"+"local" → address event,
everything else → route event. Deletions carry "deleted":true. Field
names verified against iproute2 6.x source.

ip_text parser drops the inet/inet6 requirement — any "<index>: <name>
<whitespace>..." line is an address event; no other indexed lines use
that format.

rtnetlink path now extracts AddressAttribute::Label for address events
instead of the useless "if#N" placeholder, so AddressRemoved carries
the real interface name needed by should_reconnect.
